requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
Pillow==10.4.0
openpyxl==3.1.5
starlette==0.38.5
pytest==8.3.2
requests==2.32.3
